Move contexts to their own directory
