+++ src/app/components/shell/shell.js
@@ -1,6 +1,6 @@
 import React from 'react';
-import {LanguageContextProvider} from '~/models/language-context';
-import {SubjectCategoryContextProvider} from '~/models/subject-category-context';
+import {LanguageContextProvider} from '~/contexts/language';
+import {SubjectCategoryContextProvider} from '~/contexts/subject-category';
 import {BrowserRouter} from 'react-router-dom';
 import Router from './router';
 import ReactModal from 'react-modal';
