+++ src/app/components/sticky-footer/sticky-footer.js
@@ -1,6 +1,6 @@
 import React, {useEffect} from 'react';
 import {RawHTML} from '~/components/jsx-helpers/jsx-helpers.jsx';
-import useWindowContext, {WindowContextProvider} from '~/models/window-context';
+import useWindowContext, {WindowContextProvider} from '~/contexts/window';
 import './sticky-footer.scss';
 import shellBus from '~/components/shell/shell-bus';
 import cn from 'classnames';
