+++ src/app/contexts/language.js
@@ -1,8 +1,8 @@
-import React from 'react';
+import {useReducer} from 'react';
 import buildContext from '~/components/jsx-helpers/build-context';
 
 function useContextValue() {
-    const [language, setLanguage] = React.useReducer(
+    const [language, setLanguage] = useReducer(
         (_, newValue) => {
             window.localStorage.oswebLanguage = newValue;
             return newValue;
