+++ src/app/pages/subjects/context.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import buildContext from '~/components/jsx-helpers/build-context';
 import usePageData from '~/components/jsx-helpers/page-loader';
-import useLanguageContext from '~/models/language-context';
+import useLanguageContext from '~/contexts/language';
 
 const preserveWrapping = true;
